requirements.txt
lxml==4.9.1
numpy==1.23.1
pandas==1.4.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
six==1.16.0
